app/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
mysql-connector-python
proxmoxer
requests
docker
python-dotenv
python-jose
passlib
bcrypt
pytest
httpx
pytest-cov
structlog
secure
pydantic-settings
